TASK: Move `Translation` into `model` module
