+++ packages/neos-ui-i18n/src/model/Translation.spec.ts
@@ -7,8 +7,7 @@
  * information, please view the LICENSE file which was distributed with this
  * source code.
  */
-import {Locale} from '../model';
-
+import {Locale} from './Locale';
 import {Translation} from './Translation';
 
 describe('Translation', () => {
